requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
black==24.2.0
click==8.1.7
confuse==2.0.1
flake8==7.0.0
frozenlist==1.4.1
idna==3.6
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyYAML==6.0.1
tomli==2.0.1
typing_extensions==4.10.0
yarl==1.9.4
